Issue solved with routing
