+++ remix1/app/layout/MainLayout/LogoSection/index.js
@@ -1,4 +1,3 @@
-import { Link } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 
 // material-ui
@@ -8,6 +7,7 @@ import { ButtonBase } from '@mui/material';
 import { MENU_OPEN } from '../../../store/actions';
 import config from '../../../../config';
 import Logo from '../../../ui-component/Logo';
+import { Link } from '@remix-run/react';
 
 // ==============================|| MAIN LOGO ||============================== //
