+++ remix1/app/layout/MinimalLayout/index.js
@@ -1,6 +1,6 @@
-import { Outlet } from 'react-router-dom';
 
 // project imports
+import { Outlet } from '@remix-run/react';
 import Customization from '../Customization';
 
 // ==============================|| MINIMAL LAYOUT ||============================== //
