+++ remix1/app/layout/NavigationScroll.js
@@ -1,6 +1,6 @@
+import { useLocation } from '@remix-run/react';
 import PropTypes from 'prop-types';
 import { useEffect } from 'react';
-import { useLocation } from 'react-router-dom';
 
 // ==============================|| NAVIGATION SCROLL TO TOP ||============================== //
