+++ remix1/app/routes/index.jsx
@@ -1,23 +1,9 @@
-import { CssBaseline, StyledEngineProvider, ThemeProvider } from "@mui/material";
-import { Outlet } from "@remix-run/react";
-import { useSelector } from "react-redux";
-import NavigationScroll from "../layout/NavigationScroll";
-import theme from "../themes";
-import Dashboard from "./dashboard";
-
+import Dashboard from "./dashboard/Default";
 
 export default function Index() {
-  const customization = useSelector((state) => state.customization);
   return (
     <>
-      <StyledEngineProvider injectfirst>
-        <ThemeProvider theme={theme(customization)}>
-          <CssBaseline />
-          <NavigationScroll>
-            <Dashboard/>
-          </NavigationScroll>
-        </ThemeProvider>
-      </StyledEngineProvider>
+      <Dashboard />
     </>
   );
 }
